requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
ippanel==2.0.7
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.5
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.10.6
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7